backend/requirements.txt
2# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6

# AI and ML
openai==1.59.5
anthropic==0.42.0
numpy==1.26.4

# Web Scraping and Data Collection
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0

# PDF Generation
weasyprint==63.1
reportlab==4.2.5
Pillow==11.0.0
Jinja2==3.1.4

# Email
sendgrid==6.11.0

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
